Time: 10 ms (27.75%), Space: 17.9 MB (24.72%) - LeetHub

+++ 3685-count-subarrays-of-length-three-with-a-condition/3685-count-subarrays-of-length-three-with-a-condition.py
@@ -1,8 +1,10 @@
 class Solution:
     def countSubarrays(self, nums: List[int]) -> int:
         count = 0
-        for i in range(len(nums) - 2):
-            x, y, z = nums[i], nums[i + 1], nums[i + 2]
-            if x + z == y / 2:
+        for i in range(len(nums)-2):
+            x,y,z = nums[i], nums[i+1], nums[i+2]
+
+            if float(x + z) == float(y/2):
                 count += 1
+        
         return count
